+++ gatsby-node.js
@@ -1,6 +1,6 @@
 const path = require(`path`)
 const { createFilePath } = require(`gatsby-source-filesystem`)
-
+const appRoutes = require('./app-routes')
 exports.createPages = ({ graphql, actions }) => {
   const { createPage } = actions
 
@@ -82,12 +82,50 @@ exports.onCreateNode = ({ node, actions, getNode }) => {
 
 exports.onCreatePage = async ({ page, actions }) => {
   const { createPage } = actions
-  // Only update the `/app` page.
-  if (page.path.match(/^\/app/)) {
-    // page.matchPath is a special key that's used for matching pages
-    // with corresponding routes only on the client.
-    page.matchPath = '/app/*'
-    // Update the page.
-    createPage(page)
-  }
+  const clientPages = [
+    '/app/sorting',
+    '/app/graph-search',
+    '/app/games',
+    '/app/drawable-graph',
+  ]
+  clientPages.map(name => {
+    // Only update the `/app` page.
+    if (page.path.includes(name)) {
+      // page.matchPath is a special key that's used for matching pages
+      // with corresponding routes only on the client.
+      page.matchPath = `${name}/*`
+      // Update the page.
+      createPage(page)
+    }
+  })
+}
+
+exports.sourceNodes = async ({
+  actions,
+  createContentDigest,
+  createNodeId,
+  getNodesByType,
+}) => {
+  const { createNode } = actions
+
+  // loop through data and create Gatsby nodes
+  Object.entries(appRoutes.apps).forEach(([type, item]) => {
+    Object.entries(item).forEach(([name, content]) => {
+      createNode({
+        ...content,
+        category: type,
+        name: name,
+        id: createNodeId(`${type}-${content.url}`),
+        parent: null,
+        children: [],
+        internal: {
+          type: 'app',
+          content: JSON.stringify(content),
+          contentDigest: createContentDigest(content),
+        },
+      })
+    })
+  })
+
+  return
 }