+++ src/app/containers.js
@@ -2,5 +2,4 @@ export { default as Sorting } from "./sorting";
 export { default as Graph } from "./graph-search";
 export { default as DrawableGraph } from "./drawable-graph";
 export { default as Games } from "./games";
-export { default as Home } from "./home";
 export { default as TOC } from "./theory_of_computation";