+++ src/app/games/index.js
@@ -2,13 +2,6 @@ import React, { Component } from 'react'
 import './css/index.css'
 import { Menu, Head, NavLink } from '../components'
 import { Snake } from './statefulComponents'
-// import Loadable from "react-loadable";
-import SnakeTut from './components/snake-tut'
-// const SnakeTut = Loadable({
-//   loader: () => import("./components/snake-tut"),
-//   loading: () => <div>Loading...</div>,
-//   delay: 0
-// });
 import { Router } from '@reach/router'
 
 const games = ['snake-game']
@@ -34,9 +27,9 @@ let menuItems = games.map(item => ({
 
 const data = {
   description:
-    'Play games online made using javascript, html and css. View the tutorial to make you own game.s',
-  title: 'Games | Subesh Bhandari',
-  url: '/app/games/',
+    'Play snake game online. Made using HTML CSS and javascript. Learn how to build your own snake game following the tutorial.',
+  title: 'Snake Game | Games',
+  url: '/app/games/snake-game',
 }
 
 export default class Games extends Component {