+++ src/app/home/components/index.js
@@ -1,3 +1,0 @@
-export { default as Slide } from "./slide";
-export { default as Section } from "./section";
-export { default as Features } from "./features";